requirements.txt
APScheduler==3.6.3
attrs==21.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2022.9.24
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.1.1
clarifai==2.6.2
configparser==3.8.1
cryptography==3.4.7
decorator==4.4.2
EasyProcess==0.3
emoji==1.2.0
future==0.18.2
googleapis-common-protos==1.53.0
grpcio==1.47.0
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
iniconfig==1.1.1
instabot==0.117.0
jsonschema==2.6.0
MeaningCloud-python==2.0.0
mock==4.0.3
moviepy==1.0.3
numpy==1.20.3
oauthlib==3.1.1
packaging==20.9
Pillow==9.2.0
pluggy==0.13.1
plyer==2.0.0
proglog==0.1.9
protobuf==3.17.3
py==1.10.0
pycparser==2.20
pylast==4.2.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
python-dotenv==0.17.1
python-telegram-bot==13.6
PyTumblr==0.1.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyVirtualDisplay==2.2
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.13.3
schedule==1.1.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.61.1
tweepy==3.10.0
tzdata==2022.2
tzlocal==2.1
urllib3==1.26.5
webdriverdownloader==1.1.0.3
